feat(email): Centraliza envío de correos de membresías en servicio global
Se refactoriza el módulo de membresías para utilizar el servicio de correo electrónico global (`emailService`) en lugar de su propia implementación local de `nodemailer`.

Principales cambios:
- Se mejora `emailService` para que pueda enviar correos reales utilizando `nodemailer`, manteniendo la compatibilidad hacia atrás.
- Se añade un nuevo método `sendHtmlEmail` a `emailService` para manejar contenido HTML, mientras que `sendEmail` se mantiene para texto plano.
- La lógica para generar la plantilla del comprobante de membresía se extrae a su propio archivo (`src/modules/membership/membershipEmailTemplates.js`), mejorando la modularidad.
- El servicio de correo se habilita globalmente en `src/router/routerGlobal.js`.
- Se elimina el archivo obsoleto `src/modules/membership/utils/nodeMailer.js`.
